capstone/dashboard/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.1 ; python_version >= '3.8'
aiosignal==1.3.1 ; python_version >= '3.7'
alembic==1.13.1 ; python_version >= '3.8'
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.2.0 ; python_version >= '3.8'
apache-airflow==2.8.0
apache-airflow-providers-common-sql==1.10.0 ; python_version ~= '3.8'
apache-airflow-providers-ftp==3.7.0 ; python_version ~= '3.8'
apache-airflow-providers-http==4.8.0 ; python_version ~= '3.8'
apache-airflow-providers-imap==3.5.0 ; python_version ~= '3.8'
apache-airflow-providers-sqlite==3.7.0 ; python_version ~= '3.8'
apispec[yaml]==6.3.1 ; python_version >= '3.7'
argcomplete==3.2.1 ; python_version >= '3.8'
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version < '3.11'
attrs==23.2.0 ; python_version >= '3.7'
babel==2.14.0 ; python_version >= '3.7'
backoff==2.2.1 ; python_version >= '3.7'
blinker==1.7.0 ; python_version >= '3.8'
cachelib==0.9.0 ; python_version >= '3.7'
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==4.8.0
configupdater==3.2 ; python_version >= '3.6'
connexion[flask]==2.14.2 ; python_version >= '3.6'
cron-descriptor==1.4.0
croniter==2.0.1 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==41.0.7 ; python_version >= '3.7'
deprecated==1.2.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.7 ; python_version >= '3.7'
django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-csv==3.0.2
dnspython==2.4.2 ; python_version >= '3.8' and python_version < '4.0'
docutils==0.20.1 ; python_version >= '3.7'
drf-excel==2.4.0
email-validator==1.3.1 ; python_version >= '3.5'
et-xmlfile==1.1.0 ; python_version >= '3.6'
exceptiongroup==1.2.0 ; python_version < '3.11'
fiona==1.9.5 ; python_version >= '3.7'
flask==2.2.5 ; python_version >= '3.7'
flask-appbuilder==4.3.10 ; python_version ~= '3.7'
flask-babel==2.0.0
flask-caching==2.1.0 ; python_version >= '3.7'
flask-jwt-extended==4.6.0 ; python_version >= '3.7' and python_version < '4'
flask-limiter==3.5.0 ; python_version >= '3.7'
flask-login==0.6.3 ; python_version >= '3.7'
flask-session==0.5.0 ; python_version >= '3.7'
flask-sqlalchemy==2.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-wtf==1.2.1 ; python_version >= '3.8'
frozenlist==1.4.1 ; python_version >= '3.8'
fsspec==2023.12.2 ; python_version >= '3.8'
geographiclib==2.0 ; python_version >= '3.7'
geopandas==0.14.2
geopy==2.4.1
google-re2==1.1 ; python_version ~= '3.8'
googleapis-common-protos==1.62.0 ; python_version >= '3.7'
graphviz==0.20.1 ; python_version >= '3.7'
greenlet==3.0.3 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.60.0 ; python_version >= '3.7'
gunicorn==21.2.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.2 ; python_version >= '3.8'
httpx==0.26.0
idna==3.6 ; python_version >= '3.5'
importlib-metadata==6.11.0 ; python_version >= '3.8'
importlib-resources==6.1.1 ; python_version >= '3.8'
inflection==0.5.1 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.20.0 ; python_version >= '3.8'
jsonschema-specifications==2023.12.1 ; python_version >= '3.8'
lazy-object-proxy==1.10.0 ; python_version >= '3.8'
limits==3.7.0 ; python_version >= '3.7'
linkify-it-py==2.0.2 ; python_version >= '3.7'
lockfile==0.12.2
mako==1.3.0 ; python_version >= '3.8'
markdown==3.5.1 ; python_version >= '3.8'
markdown-it-py==3.0.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
marshmallow==3.20.1 ; python_version >= '3.8'
marshmallow-oneofschema==3.0.1 ; python_version >= '3.6'
marshmallow-sqlalchemy==0.26.1 ; python_version >= '3.6'
mdit-py-plugins==0.4.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.26.3 ; python_version < '3.11'
openpyxl==3.1.2 ; python_version >= '3.6'
opentelemetry-api==1.22.0 ; python_version >= '3.7'
opentelemetry-exporter-otlp==1.22.0 ; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-common==1.22.0 ; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-grpc==1.22.0 ; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-http==1.22.0 ; python_version >= '3.7'
opentelemetry-proto==1.22.0 ; python_version >= '3.7'
opentelemetry-sdk==1.22.0 ; python_version >= '3.7'
opentelemetry-semantic-conventions==0.43b0 ; python_version >= '3.7'
ordered-set==4.1.0 ; python_version >= '3.7'
packaging==23.2 ; python_version >= '3.7'
pandas==2.1.4
pathspec==0.12.1 ; python_version >= '3.8'
pendulum==2.1.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pluggy==1.3.0 ; python_version >= '3.8'
prison==0.2.1
protobuf==4.25.1 ; python_version >= '3.8'
psutil==5.9.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3 ; python_version >= '3.7'
pydantic-core==2.14.6 ; python_version >= '3.7'
pygments==2.17.2 ; python_version >= '3.7'
pyjwt==2.8.0 ; python_version >= '3.7'
pyproj==3.6.1 ; python_version >= '3.9'
python-daemon==3.0.1 ; python_version >= '3'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-nvd3==0.15.0
python-slugify==8.0.1 ; python_version >= '3.7'
pytz==2023.3.post1
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.1 ; python_version >= '3.6'
referencing==0.32.1 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
requests-toolbelt==1.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3339-validator==0.1.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.7.0 ; python_full_version >= '3.7.0'
rich-argparse==1.4.0 ; python_version >= '3.7'
rpds-py==0.16.2 ; python_version >= '3.8'
setproctitle==1.3.3 ; python_version >= '3.7'
setuptools==69.0.3 ; python_version >= '3.8'
shapely==2.0.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.51
sqlalchemy-jsonfield==1.0.2 ; python_full_version >= '3.7.0'
sqlalchemy-utils==0.41.1 ; python_version >= '3.6'
sqlparse==0.4.4 ; python_version >= '3.5'
tabulate==0.9.0 ; python_version >= '3.7'
tenacity==8.2.3 ; python_version >= '3.7'
termcolor==2.4.0 ; python_version >= '3.8'
text-unidecode==1.3
typing-extensions==4.9.0 ; python_version >= '3.8'
tzdata==2023.4 ; python_version >= '2'
uc-micro-py==1.0.2 ; python_version >= '3.7'
unicodecsv==0.14.1
universal-pathlib==0.1.4 ; python_version >= '3.8'
urllib3==2.1.0 ; python_version >= '3.8'
werkzeug==2.2.3 ; python_version >= '3.7'
whitenoise==6.6.0
wrapt==1.16.0 ; python_version >= '3.6'
wtforms==3.1.2 ; python_version >= '3.8'
yarl==1.9.4 ; python_version >= '3.7'
zipp==3.17.0 ; python_version >= '3.8'
